tests/wpt/web-platform-tests/html/browsers/browsing-the-web/navigating-across-documents/source/navigate-child-function-parent.html

<!doctype html>
<meta charset=utf-8>
<title>Set location from a parent</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<iframe></iframe>
<script>
  onload = function() {
    var fr = document.querySelector("iframe")
    fr.contentWindow.location = "support/dummy.html"
    fr.onload = function() {
      assert_equals(fr.contentDocument.referrer, document.URL)
      done()
    }
  }
</script>
